Refactor actions
